refactor(modularization): Phase 5 step 5-S2a — conversation DB carve (task #26)
Physically move the three conversation-owned ORM classes — ``Bot`` /
``Chat`` / ``TurnFeedback`` — and their six enums (``BotStatus`` /
``BotType`` / ``ChatStatus`` / ``ChatPeerType`` / ``TurnFeedbackType``
/ ``TurnFeedbackTag``) into the new ``aperag/domains/conversation/db``
package, and replace the former definitions in ``aperag.db.models``
with a re-export shim so the pre-refactor caller base (~40 import
sites across services, routes, alembic env.py, and tasks) keeps
working without a rename sweep.

The ``Chat.get_bot`` / ``set_bot`` and ``TurnFeedback.get_chat`` /
``set_chat`` helper methods travel with the classes — both are
intra-domain lookups (Chat → Bot, TurnFeedback → Chat) so no
cross-domain dependency is introduced.

``_random_id()`` and ``_enum_column()`` helper twins mirror the Phase
3 pattern: each domain DB module carries its own so the G1
strict-ban aggregate stays off the import graph. Phase 6 cleanup
consolidates onto ``aperag.db.base``.

``Role`` enum stays in ``aperag.db.models`` — it is identity scope
(Phase 4) and is referenced by ``Invitation`` at class-body level
via ``Column(EnumColumn(Role), ...)``; the cleaner location for it
is the Phase 4 identity domain, which the ``chenyexuan/phase4-impl``
PR is handling in parallel.

Re-export block in ``aperag/db/models.py`` is merged into the
pre-existing Phase 3 per-domain block (alphabetised, with a unified
docstring). ``aperag.db.models`` post-5-S2a re-exports 24 symbols
(Phase 3 15 + conversation 9) — G11 extension path, locks in at 5-S8.

Canonical: msg=92f5788d Section 1 5-S2 + Section 3 target path.

Verify:
* ``uv run python -c "from aperag.db.models import Bot, BotStatus, BotType, Chat, ChatStatus, ChatPeerType, TurnFeedback, TurnFeedbackType, TurnFeedbackTag"`` OK
* ``Bot is aperag.domains.conversation.db.models.Bot`` identity preserved
* ``pytest tests/unit_test/`` 671 passed / 29 skipped
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
* ``Base.metadata`` still registers the ``bot`` / ``chat`` /
  ``turn_feedback`` tables after the move.
